Clear every git lock before an update, and stamp the APK with its commit
An installation that was interrupted while a group of files was downloading
could not be repaired: git holds .git/info/sparse-checkout.lock for the whole
group, and the sweep before an update only looked at the top level of .git, so
every later Install/Update died with "File exists" until the app's data was
cleared. It now sweeps the whole directory, and takes the half written packs of
an interrupted fetch with it.

The manifest set versionCode 1 / versionName 1.0, and aapt2 only fills its own
values in when it finds none there, so every build carried the same version no
matter which commit it came from. The manifest sets neither now, the build fails
if that ever comes back, and the workflow checks out the full history because
the version code counts commits.

Also:
- the watcher thread decided on the main thread's field whether the process that
  ended is still the current one, which could tear down a server a network
  change had just restarted
- a burst of installer output queued one full redraw of the screen per line
- an xz dictionary size of 2 GB or more overflowed to a negative int and slipped
  past the check meant to reject it
- a package entry replacing a link whose target is gone was written through the
  link, and a path leading out of the prefix through one is now refused
- the Android workflow fails a pull request that changes the app without
  refreshing the committed APK

The new installer test suite clones a repository it builds itself with the
machine's git: a first installation, an update, and a resume out of an
interrupted download whose lock has to be cleared.

+++ .android/src/io/virtualtabletop/server/Xz.java
@@ -163,11 +163,13 @@ final class Xz extends InputStream {
     int properties = header[position[0]] & 0xFF;
     if(properties > 40)
       throw new IOException("unsupported dictionary size");
-    int size = properties == 40 ? MAXIMUM_DICTIONARY : (2 | (properties & 1)) << (properties / 2 + 11);
+    // in long, because the sizes the last few property values stand for do not fit into an int
+    // and would come out negative, which is a size that passes every check below it
+    long size = properties == 40 ? MAXIMUM_DICTIONARY : (long)(2 | (properties & 1)) << (properties / 2 + 11);
     if(size > MAXIMUM_DICTIONARY)
       throw new IOException("dictionary of " + size + " bytes is too large");
     if(dictionary.length < size)
-      dictionary = new byte[size];
+      dictionary = new byte[(int)size];
 
     resetDictionary();
     chunkRemaining = 0;
